requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
flask-mysqldb
tensorflow-cpu
pillow
numpy
gunicorn
flask-cors
python-dotenv
flask-migrate
psycopg2-binary